+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/CsvResponseWriter.java
@@ -4,6 +4,7 @@ package com.yahoo.bard.webservice.web;
 
 import com.yahoo.bard.webservice.application.ObjectMappersSuite;
 import com.yahoo.bard.webservice.data.metric.MetricColumn;
+import com.yahoo.bard.webservice.web.apirequest.ApiRequest;
 
 import com.fasterxml.jackson.databind.ObjectWriter;
 import com.fasterxml.jackson.dataformat.csv.CsvMapper;
