+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/FiliResponseWriter.java
@@ -2,6 +2,8 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.web;
 
+import com.yahoo.bard.webservice.web.apirequest.ApiRequest;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
