+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/FiliResponseWriterSelector.java
@@ -3,6 +3,8 @@
 package com.yahoo.bard.webservice.web;
 
 
+import com.yahoo.bard.webservice.web.apirequest.ApiRequest;
+
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
