+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/JsonApiResponseWriter.java
@@ -7,6 +7,7 @@ import com.yahoo.bard.webservice.data.Result;
 import com.yahoo.bard.webservice.data.dimension.Dimension;
 import com.yahoo.bard.webservice.data.dimension.DimensionColumn;
 import com.yahoo.bard.webservice.data.dimension.DimensionField;
+import com.yahoo.bard.webservice.web.apirequest.ApiRequest;
 
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonGenerator;
