+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/MetricsApiRequest.java
@@ -3,6 +3,7 @@
 package com.yahoo.bard.webservice.web;
 
 import com.yahoo.bard.webservice.data.metric.LogicalMetric;
+import com.yahoo.bard.webservice.web.apirequest.ApiRequest;
 
 import java.util.Set;
