+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/RequestMapper.java
@@ -3,6 +3,7 @@
 package com.yahoo.bard.webservice.web;
 
 import com.yahoo.bard.webservice.data.config.ResourceDictionaries;
+import com.yahoo.bard.webservice.web.apirequest.ApiRequest;
 
 import javax.validation.constraints.NotNull;
 import javax.ws.rs.container.ContainerRequestContext;
