+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/ResponseWriterSelector.java
@@ -3,6 +3,8 @@
 package com.yahoo.bard.webservice.web;
 
 
+import com.yahoo.bard.webservice.web.apirequest.ApiRequest;
+
 import java.util.Optional;
 
 /**
