+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/ResponseWriterSelectorSpec.groovy
@@ -9,6 +9,7 @@ import com.yahoo.bard.webservice.data.ResultSet
 import com.yahoo.bard.webservice.table.Column
 import com.yahoo.bard.webservice.util.Pagination
 import com.yahoo.bard.webservice.util.SimplifiedIntervalList
+import com.yahoo.bard.webservice.web.apirequest.ApiRequest
 
 import org.joda.time.DateTime
