+++ fili-security/src/test/groovy/com/yahoo/bard/webservice/web/security/DataApiRequestMapperUtilsSpec.groovy
@@ -3,7 +3,7 @@
 package com.yahoo.bard.webservice.web.security
 
 import com.yahoo.bard.webservice.data.config.ResourceDictionaries
-import com.yahoo.bard.webservice.web.ApiRequest
+import com.yahoo.bard.webservice.web.apirequest.ApiRequest
 import com.yahoo.bard.webservice.web.RequestMapper
 import com.yahoo.bard.webservice.web.RequestValidationException
